Initialize the Git repository
The initial commit of this repository. In fact, I had the idea of
creating a Git repository when I started the project, but I seemed to
forget about it. Now with the license chosen, I added the license
headers and initialized the Git repository, although it might be a pity
that some history of the project got lost.

Signed-off-by: xianfengting <xianfengting221@163.com>